Extract CallerInformation and ArgumentExtractor to module-level files
Both were nested inside Instrumenter, making them inaccessible outside and
tightly coupling them to Instrumenter's implementation. Moving them to their
own files (caller_information.rb, argument_extractor.rb) under Observable::
makes them independently testable and reusable.

Also takes the opportunity to fix the unused local_variables call and bare
rescue in ArgumentExtractor (fixes already done in sibling PRs).
