+++ lib/observable/caller_information.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Observable
+  CallerInformation = Struct.new(:method_name, :namespace, :filepath, :line_number, :arguments, :is_class_method, keyword_init: true)
+end
